Fixed a bug in train.py
The output of the encoder-decoder model ranges from 0 to 1 because of the sigmoid activation function. The mask or labels to calculate the cost function should also have the same range. I tried your code and it was not giving me the desired results. This small change has a great effect on the training time, accuracy and output.